+++ src/react-components/src/__tests__/components/Heading-test.tsx
@@ -15,14 +15,26 @@ describe("Section Headings", () => {
   });
   it("Throws error when invalid heading number passed", () => {
     expect(() => Enzyme.mount(<Heading id="h1" level={9} text={"Heading 9"} />))
-    .to.throw("Heading only supports levels 1-6");
+      .to.throw("Heading only supports levels 1-6");
   });
   it("Can pass heading content as child", () => {
     wrapper = Enzyme.shallow(<Heading id="h1" level={1}>Text</Heading>);
     expect(wrapper.find("h1")).to.have.lengthOf(1);
   });
-  it("Throws error when invalid heading is passed as child", () => {
-    expect(() => Enzyme.mount(<Heading id="h1" level={3}><span>oh no</span></Heading>))
-    .to.throw("Headings can only be plain text or bold");
+  it("Can pass heading content as child span", () => {
+    wrapper = Enzyme.shallow(<Heading id="h1" level={1}>Text <span>hi</span></Heading>);
+    expect(wrapper.find("h1")).to.have.lengthOf(1);
   });
+  it("Has <a> tag when passed URL", () => {
+    wrapper = Enzyme.shallow(<Heading id="h1" level={1} url="fake-url" text={"Heading 1"} />);
+
+    expect(wrapper.find("h1")).to.have.lengthOf(1);
+    expect(wrapper.find("a")).to.have.lengthOf(1);
+  });
+  it("<a> has class when passed urlClass", () => {
+    wrapper = Enzyme.shallow(<Heading id="h1" level={1} url="fake-url" urlClass="fake-class" text={"Heading 1"} />);
+    expect(wrapper.find("h1")).to.have.lengthOf(1);
+    expect(wrapper.find("a").hasClass("fake-class")).to.equal(true);
+  }
+  );
 });