+++ src/react-components/src/components/01-atoms/Text/Headings/Heading.stories.tsx
@@ -17,4 +17,17 @@ export const allHeadings = () => [
   <Heading level={6} id="heading6" text={"Heading 6"} />
 ];
 
-export const headingsWithBold = () => <Heading id="heading1" level={1}><b>Head</b>ing 1</Heading> ;
+export const headingsWithBold = () => <Heading id="heading1" level={1}><b>Head</b>ing 1</Heading>;
+export const headingsWithLink = () => <Heading id="heading1" level={1}><a href="hello">Head</a>ing</Heading>;
+
+export const pageTitleAT79 = () => <Heading level={1} id="hello-page-title"
+  blockName="page-title"
+  text="hello">
+</Heading>;
+
+export const researchNowSectionTitle = () => <Heading level={1}
+  id="researchNow-page-title-id"
+  baseClass="rn-section-title"
+  url="rn-homepage-url"
+  urlClass="rn-section-title"> Research<span className={"rn-section-title__emphasis"}>Now
+</span></Heading>;