+++ src/react-components/src/components/03-organisms/SearchPromo/SearchPromo.tsx
@@ -27,7 +27,9 @@ export interface SearchPromoProps {
   textChangeHandler?: (event: React.FormEvent) => void;
 }
 export default function SearchPromo(props: SearchPromoProps) {
-  const { headingText, modifiers, blockName, titleId, searchBarId, searchButtonId, hasError, errorMessage,
+  const { headingText, modifiers, blockName, titleId,
+    searchBarId, searchButtonId,
+    hasError, errorMessage,
     dropdownId, selectedOption, searchDropdownOptions, advancedSearchMessage, searchValue, searchInputAriaLabel,
     selectChangeHandler, selectBlurHandler, searchSubmitHandler, textChangeHandler } = props;
 
@@ -40,12 +42,13 @@ export default function SearchPromo(props: SearchPromoProps) {
       <div className={bem("search-section", [], searchpromo__base_class)}>
         <SearchBar
           searchBarId={searchBarId}
+          searchBarAriaLabelledBy={titleId}
           buttonId={searchButtonId}
           dropdownId={dropdownId}
           dropdownOptions={searchDropdownOptions}
+          dropdownAriaLabel="Filter Search"
           selectedField={selectedOption}
           searchValue={searchValue}
-          textFieldAriaLabel={searchInputAriaLabel}
           hasError={hasError}
           errorMessage={errorMessage}
           selectBlurHandler={selectBlurHandler}